(#50) add the CRYPTEX_FOLDER env variable to set a box folder other than the default one
