+++ cmd/edit/common.go
@@ -1,0 +1,6 @@
+package edit
+
+var (
+	verbose bool
+	boxName string
+)
